+++ src/components/Spinner/index.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { CircularProgress } from '@material-ui/core';
+
+const Spinner = () => {
+  return (
+    <div
+      style={{
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}
+    >
+      <CircularProgress
+        size={55}
+        style={{ margin: '0 auto' }}
+        color="inherit"
+      />
+    </div>
+  );
+};
+
+export default Spinner;