+++ src/components/index.js
@@ -1,1 +1,3 @@
 export { default as CustomSwitch } from './CustomSwitch';
+export { default as CustomInput } from './CustomInput';
+export { default as Spinner } from './Spinner';