+++ src/layouts/Public/index.js
@@ -1,16 +1,28 @@
 import React from 'react';
+
+import { makeStyles } from '@material-ui/core';
+
 import Footer from './components/Footer';
 import Header from './components/Header';
 
 const headerHeight = 140;
 
+const useStyles = makeStyles((theme) => ({
+  appWrapper: {},
+  children: {
+    backgroundColor: theme.palette.white,
+  },
+}));
+
 const PublicLayout = ({ children }) => {
+  const classes = useStyles();
+
   return (
-    <React.Fragment>
+    <div className={classes.appWrapper}>
       <Header height={headerHeight} />
-      {children}
+      <div className={classes.children}>{children}</div>
       <Footer />
-    </React.Fragment>
+    </div>
   );
 };