+++ src/views/Home/index.js
@@ -1,17 +1,16 @@
 import React, { useState } from 'react';
 
+import { useHistory } from 'react-router-dom';
 import {
-  Paper,
-  InputBase,
   Typography,
-  IconButton,
   Button,
   FormControlLabel,
   makeStyles,
 } from '@material-ui/core';
+
 import SearchIcon from '@material-ui/icons/Search';
 
-import { CustomSwitch } from 'components';
+import { CustomSwitch, CustomInput } from 'components';
 
 const useStyles = makeStyles((theme) => ({
   mainWrapper: {
@@ -19,8 +18,8 @@ const useStyles = makeStyles((theme) => ({
     justifyContent: 'center',
     alignItems: 'center',
     flexDirection: 'column',
-    padding: '100px 0',
-    backgroundColor: theme.palette.white,
+    paddingBottom: 150,
+    paddingTop: 150,
   },
   mainContent: {
     maxWidth: 800,
@@ -28,30 +27,14 @@ const useStyles = makeStyles((theme) => ({
   textSection: {},
   mainTitle: {
     fontSize: '4em',
-    color: theme.palette.black,
   },
   secondaryTitle: {
     fontSize: '1.5em',
-    color: theme.palette.black,
-    marginTop: '1em',
+    marginTop: '1rem',
   },
   searchEmailInputWrapper: {
     display: 'flex',
     margin: '2.5em 0 1.5em 0',
-  },
-  searchEmailInputPaper: {
-    padding: '2px 4px',
-    display: 'flex',
-    alignItems: 'center',
-    width: '100%',
-  },
-  searchIcon: {
-    padding: 10,
-  },
-  searchEmailInput: {
-    marginLeft: theme.spacing(1),
-    flex: 1,
-    fontSize: '2em',
   },
   searchButton: {
     color: theme.palette.white,
@@ -63,10 +46,13 @@ const useStyles = makeStyles((theme) => ({
 
 const Home = () => {
   const classes = useStyles();
+  const history = useHistory();
   const [switchDNIToCode, setSwitchDNIToCode] = useState(false);
 
   const handleSwitchDNIToCode = () => setSwitchDNIToCode(!switchDNIToCode);
-
+  const validateCredentials = () => {
+    history.push('/validate-credentials');
+  };
   return (
     <React.Fragment>
       <main className={classes.mainWrapper}>
@@ -82,37 +68,22 @@ const Home = () => {
             </Typography>
           </div>
           <div className={classes.searchEmailInputWrapper}>
-            <Paper
-              component="form"
-              className={classes.searchEmailInputPaper}
-              elevation={0}
-            >
-              <IconButton
-                disabled
-                type="submit"
-                className={classes.searchIcon}
-                aria-label="search-icon"
-              >
-                <SearchIcon />
-              </IconButton>
-              {/* TODO: validate the DNI, just 8 characters */}
-              {/* TODO: validate the codeUNI, just 9 characters */}
-              <InputBase
-                spellCheck="false"
-                autoFocus
-                type="text"
-                className={classes.searchEmailInput}
-                placeholder={
-                  switchDNIToCode ? 'Ingresa tu código UNI' : 'Ingresa tu DNI'
-                }
-                inputProps={{ 'aria-label': 'verify email' }}
-              />
-            </Paper>
+            <CustomInput
+              icon={<SearchIcon />}
+              spellCheck="false"
+              autoFocus
+              type="text"
+              placeholder={
+                switchDNIToCode ? 'Ingresa tu código UNI' : 'Ingresa tu DNI'
+              }
+              inputProps={{ 'aria-label': 'verify email' }}
+            />
             <Button
               size="large"
               variant="contained"
               color="primary"
               className={classes.searchButton}
+              onClick={validateCredentials}
             >
               Buscar
             </Button>