+++ src/views/index.js
@@ -1,3 +1,4 @@
 export { default as Home } from './Home';
+export { default as ValidateCredentials } from './ValidateCredentials';
 export { default as Error401 } from './Errors/401';
 export { default as Error404 } from './Errors/404';